refactor: 파일 구조를 수정한다.
