--- src/storages/ChangesStorage.js
@@ -1,3 +1,0 @@
-import Storage from '../core/Storage.js';
-
-export default new Storage('changes');
