--- src/storages/UserStorage.js
@@ -1,3 +1,0 @@
-import Storage from '../core/Storage.js';
-
-export default new Storage('user-information');
