+++ src/storages/index.js
@@ -1,0 +1,5 @@
+import Storage from '../core/Storage.js';
+
+export const ChangesStorage = new Storage('changes');
+export const ProductsStorage = new Storage('products');
+export const UserStorage = new Storage('user');
